requirements.txt
werkzeug==2.3.7
Flask
flask_sqlalchemy
flask_login
flask_wtf
WTForms
python-dotenv
email_validator
gunicorn
flask
googlemaps
SQLAlchemy
pymysql
mysql-connector-python
mysqlclient
SQLAlchemy
pymysql
mysql-connector-python
mysqlclient
bcrypt
pytest